requirements.txt
asgiref==3.7.2 
celery==5.2.7
django==4.2.2
djangorestframework==3.14.0
fastapi==0.104.1
gunicorn==21.2.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.3.4
requests==2.31.0
uvicorn==0.24.0.post1